UNTESTED: Retrieve multiple Steam libraries
The previous iteration was looking in SteamPath/config/config.vdf for multiple libraries on the computer; however, this information is only stored on my machine in SteamPath/steamapps/libraryfolders.vdf. I've just duplicated the lookup path, and renamed the variables appropriately.

Also, minor grammatical fixes and replacements of Uplay with Steam to comments and logs.